[S7] feat: pipeline de features ML en cycle_repo (Issue #41)
