+++ backend/app/repositories/__init__.py
@@ -6,6 +6,7 @@ from app.repositories.cycle_repo import (
     update_cycle,
     delete_cycle,
     count_cycles,
+    get_cycles_with_features,
 )
 from app.repositories.daily_log_repo import (
     get_logs_by_cycle,
@@ -40,10 +41,6 @@ from app.repositories.insight_repo import (
 from app.repositories.ml_repo import (
     upsert_ml_model,
     get_ml_model,
-)
-
-from app.repositories.cycle_features_repo import (
-    get_cycles_with_features,
 )
 
 __all__ = [
